pre-compile contracts before running tests
